+++ lib/domain/record/components/supports/record_page_pill_option.dart
